pyproject.toml
[project]
name = "salesgym"
version = "0.1.0"
description = "Co-evolutionary self-improving call center agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "anthropic>=0.49.0",
    "google-genai>=1.0.0",
    "elevenlabs>=1.0.0",
    "streamlit>=1.41.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]
